requirements.txt
# Dubai RE Monitor — Python dependencies
# Install: pip install -r requirements.txt

# Core data
pandas>=2.1.0
numpy>=1.26.0

# Database
sqlalchemy>=2.0.0

# Web / scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Dashboard
streamlit>=1.32.0
plotly>=5.19.0

# Scheduler (Windows-compatible, no cron needed)
APScheduler>=3.10.0

# Alerts
twilio>=8.11.0

# Config
python-dotenv>=1.0.0
